# v0/openapi/aihub_openapi.yaml
openapi: 3.1.0
info:
  title: ieops-aihub
  version: 0.4.0
  description: |
    polyforge-v3 backend. v3-design.md §5 (schema) + §10.1 (MCP tools)。
    所有 mutating endpoint 强制带 AttemptCredential 三联 (attempt_id, claim_epoch, session_secret)。
servers:
  - url: http://10.146.0.16
    description: GMI 内网默认部署 (design §6.1 推荐 HTTPS for 公网, warn-only on http://非localhost)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: aihub_sk_*
  responses:
    Unauthorized:
      description: 401 — invalid api_key / session_secret / lease
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    Conflict:
      description: 409 — epoch mismatch / lease expired / lock hard_block / busy_not_takeover_eligible
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
    AttemptCredential:
      type: object
      required: [attempt_id, claim_epoch, session_secret]
      properties:
        attempt_id: {type: string, pattern: "^ra_[a-z0-9]+$"}
        claim_epoch: {type: integer, minimum: 1}
        session_secret:
          type: string
          pattern: "^[0-9a-f]{64}$"
          minLength: 64
          maxLength: 64
          description: "raw 32-byte hex (64 chars); server sha256 比对存储的 session_secret_hash (§6.1)"
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - CONFLICT_EPOCH_MISMATCH
            - CONFLICT_LEASE_EXPIRED
            - CONFLICT_LEASE_TAKEN_OVER
            - CONFLICT_HARD_BLOCK
            - CONFLICT_BUSY_NOT_TAKEOVER_ELIGIBLE
            - CONFLICT_DUPLICATE_REQUEST
            - BAD_REQUEST
            - NOT_FOUND
            - PAYLOAD_TOO_LARGE
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - UNKNOWN_REMOTE_ERROR
          description: "13 known ErrorCode (§6.1) + UNKNOWN_REMOTE_ERROR forward-compat placeholder; SoT polyforge_v3.aihub.errors.ErrorCode"
        message: {type: string}
        details: {type: object, additionalProperties: true}
...
    ArtifactActionRequest:
      allOf:
        - $ref: '#/components/schemas/AttemptCredential'
        - type: object
          required: [type, identifier, repo]
          properties:
            type: {type: string, enum: [pr, branch, issue]}
            identifier: {type: string}
            repo: {type: string}
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: {type: boolean}
...
paths:
...
  /v1/work_items/{work_item_id}/artifacts/adopt:
    parameters:
      - {name: work_item_id, in: path, required: true, schema: {type: string}}
    post:
      operationId: adopt_artifact
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactActionRequest'}}}
      responses:
        '200':
          description: adopted
          content: {application/json: {schema: {$ref: '#/components/schemas/OkResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/work_items/{work_item_id}/artifacts/ignore:
    parameters:
      - {name: work_item_id, in: path, required: true, schema: {type: string}}
    post:
      operationId: ignore_artifact
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactActionRequest'}}}
      responses:
        '200':
          description: ignored
          content: {application/json: {schema: {$ref: '#/components/schemas/OkResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/work_items/{work_item_id}/artifacts/close:
    parameters:
      - {name: work_item_id, in: path, required: true, schema: {type: string}}
    post:
      operationId: close_artifact
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactActionRequest'}}}
      responses:
        '200':
          description: closed
          content: {application/json: {schema: {$ref: '#/components/schemas/OkResponse'}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
...